#216 Add match to mockUpdate via inherited mock-request-match
